added PDB.read mmtf tests to test_mmtf and fix doctest
